* kparse.c (strutol): Cast argument to isupper()/tolower() to int. 	* realmofhost.c (krb_realmofhost): Likewise.
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13369 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb4/kparse.c
@@ -562,8 +562,8 @@ static char * strutol( start )
 {
     char *q;
     for (q=start; *q; q++)
-        if (isupper(*q))
-	    *q=tolower(*q);
+        if (isupper((int) *q))
+	    *q=tolower((int) *q);
     return(start);
 }
